fix(symbc): preserve long width in LCMP comparison
The greater-than branch of the symbolic LCMP cast its long operands to int before adding the path constraint, truncating values outside the 32-bit range and producing incorrect constraints for long comparisons. Pass the operands as long, matching the less-than and equal branches. Add a symbolic-execution test comparing a long beyond Integer.MAX_VALUE.

+++ jpf-symbc/src/main/gov/nasa/jpf/symbc/bytecode/LCMP.java
@@ -118,9 +118,9 @@ public class LCMP extends gov.nasa.jpf.jvm.bytecode.LCMP {
                     if (sym_v2 != null) { // both are symbolic values
                         pc._addDet(Comparator.GT, sym_v2, sym_v1);
                     } else
-                        pc._addDet(Comparator.GT, (int) v2, sym_v1);
+                        pc._addDet(Comparator.GT, v2, sym_v1); // v2 is long; (int) cast truncates values outside 32-bit range
                 } else
-                    pc._addDet(Comparator.GT, sym_v2, (int) v1);
+                    pc._addDet(Comparator.GT, sym_v2, v1); // v1 is long; same — matches LT branch at lines 94/96
                 if (!pc.simplify()) {// not satisfiable
                     th.getVM().getSystemState().setIgnored(true);
                 } else {
